requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot>=20.0
SQLAlchemy
psycopg2-binary
APScheduler
pydantic
python-dotenv
